Update setup.py
To avoid the following deprecation: `DEPRECATION: fuzzydata 0.0.8 has a non-standard dependency specifier SQLAlchemy>=1.4.31<2.0.0. pip 24.0 will enforce this behaviour change. A possible replacement is to upgrade to a newer version of fuzzydata or contact the author to suggest that they release a version with a conforming dependency specifiers. Discussion can be found at https://github.com/pypa/pip/issues/12063`